requirements.txt
sounddevice
scipy
requests
python-dotenv
ctransformers
llama-cpp-python
openai 